pyproject.toml
[tool.poetry]
name = "dashboard-release-notes"
version = "0.1.0"
description = "Script to generate release notes by extracting GitHub PR descriptions for commit hashes"
authors = ["Jonatan Witoszek"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8.1"
requests = "^2.31.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
isort = "^5.12.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
release-notes = "dashboard_release_notes.main:main"
